backend/requirements.txt

asgiref==3.5.0
backports.zoneinfo==0.2.1
breezy==3.0.2
certifi==2019.11.28
chardet==3.0.4
configobj==5.0.6
dbus-python==1.2.16
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
dulwich==0.19.15
factory-boy==3.2.1
Faker==12.1.0
fastimport==0.9.8
idna==2.8
importlib-metadata==4.10.1
Markdown==3.3.6
PyGObject==3.36.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
sqlparse==0.4.2
urllib3==1.25.8
zipp==3.7.0